🤖 AI Fix: Resolved 59 issues identified by analyzer

+++ ft_putnbr_fd.c
@@ -19,7 +19,7 @@ void	ft_putnbr_fd(int n, int fd)
 	else if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
-		ft_putnbr_fd(-n, fd);
+        ft_putnbr_fd((long)-n, fd);
 	}
 	else if (n >= 10)
 	{
